Se implementan check de errores de characters service

+++ src/main/java/com/alkemy/disney/service/impl/CharactersService.java
@@ -6,6 +6,8 @@ import com.alkemy.disney.dto.Characters.ListCharactersDTO;
 import com.alkemy.disney.dto.Characters.PostCharactersDTO;
 import com.alkemy.disney.entity.CharacterDat;
 import com.alkemy.disney.entity.Film;
+import com.alkemy.disney.exception.DatabaseError;
+import com.alkemy.disney.exception.ServiceError;
 import com.alkemy.disney.mapper.CharacterMapper;
 import com.alkemy.disney.repository.CharacterDatRepository;
 import com.alkemy.disney.service.CharactersServiceInterface;
@@ -31,32 +33,43 @@ public class CharactersService implements CharactersServiceInterface {
     }
 
     @Override
-    public CharactersDTO getById(long id) {
+    public CharactersDTO getById(long id) throws DatabaseError {
+        if(characterDatRepository.existsById(id))
         return characterMapper.charactersToDTO(characterDatRepository.getById(id));
+        else
+            throw new DatabaseError("No se pudo encontrar un personaje con ese id");
     }
 
     @Override
-    public CharactersDTO newCharacter(PostCharactersDTO newChara) {
-        CharacterDat newCharacter = characterMapper.PostCharactersDToCharacterDat(newChara);
-        characterDatRepository.save(newCharacter);
-        return characterMapper.charactersToDTO(newCharacter);
-    }
-
-    @Override
-    public void deleteCharacter(Long delcharId) {
-        CharacterDat characterDat = characterDatRepository.getById(delcharId);
-        for(Film film : characterDat.getActFilm()){
-            film.getCharacters().remove(characterDat);
+    public CharactersDTO newCharacter(PostCharactersDTO newChara) throws ServiceError {
+        if(newChara.getId()!=null&&newChara.getAge()!=null&&newChara.getName()!=null&&newChara.getStory()!=null) {
+            CharacterDat newCharacter = characterMapper.PostCharactersDToCharacterDat(newChara);
+            characterDatRepository.save(newCharacter);
+            return characterMapper.charactersToDTO(newCharacter);
         }
-        characterDatRepository.deleteById(delcharId);
+        throw new ServiceError("Los campos deben ser no nulos");
     }
 
     @Override
-    public CharactersDTO editCharacter(PostCharactersDTO editCharacter) {
-        CharacterDat characterDat = characterMapper.PostCharactersDToCharacterDat(editCharacter);
-        if(characterDatRepository.findById(characterDat.getId()).isEmpty())
-            return null;
-        characterDatRepository.save(characterDat);
-        return characterMapper.charactersToDTO(characterDat);
+    public void deleteCharacter(Long delcharId) throws DatabaseError {
+        if(characterDatRepository.existsById(delcharId)) {
+            CharacterDat characterDat = characterDatRepository.getById(delcharId);
+            for (Film film : characterDat.getActFilm()) {
+                film.getCharacters().remove(characterDat);
+            }
+            characterDatRepository.deleteById(delcharId);
+        }else
+            throw new DatabaseError("No se pudo encontrar un personaje con ese id");
+    }
+
+    @Override
+    public CharactersDTO editCharacter(PostCharactersDTO editCharacter) throws DatabaseError {
+        if(characterDatRepository.existsById(editCharacter.getId())) {
+            CharacterDat characterDat = characterMapper.PostCharactersDToCharacterDat(editCharacter);
+            characterDatRepository.save(characterDat);
+            return characterMapper.charactersToDTO(characterDat);
+        }
+        else
+            throw new DatabaseError("No se pudo encontrar un personaje con ese id");
     }
 }
